requirements.txt
office365-rest-python-client
tqdm
colorama
pyfiglet
rich